backend/engine/live_update/requirements.txt
gevent
motor


python-socketio = {extras = ["client"], version = "*"}
socketIO-client

python-dotenv
flask
websockets
aiohttp